Add tester script
The testing approach here was to generate multiple random files and give them as input to the algorithm.
The tests show that the algorithm may fail to reconstruct the correct file sometimes. This is expected, as in this version of the code we are only using the Rolling Hashes for comparing the signatures, and it's very possible for them to collide.

Signed-off-by: Matheus Cardoso <dacachmatheus@gmail.com>