Fix algorithm by comparing strong hashes when rolling hashes match
When comparing two blocks, it may be the case that the rolling hashes match, but the contents are not equal (rolling hash collision). When that happens, we compute and compare the strong hashes now. If the contents are equal, the strong hashes will match too.

Note that there is still a chance for the algorithm to fail.
If both rolling hash and strong hash collide, the reconstructed file will be wrong. But the chances of such a thing happening is astronomically low.

Signed-off-by: Matheus Cardoso <dacachmatheus@gmail.com>